@@ -3,7 +3,7 @@ import re
 from dataclasses import dataclass
 from logging import getLogger
 from pathlib import Path
-from typing import Callable, Generic, Type, TypeAlias, TypeVar
+from typing import Callable, Type, TypeAlias
 
 import jinja2
 
@@ -23,7 +23,6 @@ logger = getLogger(__name__)
 PythonCode: TypeAlias = str
 GeneratedPythonCode: TypeAlias = str
 
-T = TypeVar("T")
 
 AUTO_GENERATED_DOC_COMMENT_PATTERN = re.compile(
     r"# Code generated by restrun \".+\"\.\n"
@@ -73,7 +72,7 @@ def is_auto_generated_or_empty(source: Path | PythonCode) -> bool:
 
 
 @dataclass(frozen=True)
-class ClassInfo(Generic[T]):
+class ClassInfo[T]:
     module_name: str
     class_name: str
     class_type: Type[T]
