@@ -1,7 +1,5 @@
-from typing import Any, TypeVar
+from typing import Any
 
-
-T = TypeVar("T")
 
 try:
     from pydantic import TypeAdapter  # type: ignore
@@ -10,9 +8,8 @@ try:
 
 
 except ImportError:  # pragma: no cover
-    from typing import Generic
 
-    class TypeAdapter(Generic[T]):
+    class TypeAdapter[T]:
         def __init__(self, __type: type[T], *args, **kwargs) -> None:
             pass
 
@@ -22,7 +19,7 @@ except ImportError:  # pragma: no cover
     _USE_PYDANTIC = False
 
 
-def validate_schema(data: Any, adapter: TypeAdapter[T]) -> T:
+def validate_schema[T](data: Any, adapter: TypeAdapter[T]) -> T:
     if _USE_PYDANTIC:
         return adapter.validate_python(data)
 
